// src/features/tests/types.ts
export type WithTwoElementsProps = {
  value1: number;
  value2: number;
  expectedValue1: number;
  expectedValue2: number;
  mergedIndex: number | false;
};

export type WithThreeElementsProps = {
  value1: number;
  value2: number;
  value3: number;
  expectedValue1: number;
  expectedValue2: number;
  expectedValue3: number;
  mergedIndex: number | false;
};